test: Moved all xEventTriggered tests into separate integrationTest suite
